Refactor structure to make it easier to find things
