+++ bot/app/util/commands.py
@@ -1,20 +1,18 @@
 import json
 import discord
 import subprocess
-
-import foodflex.util.data as data
-import foodflex.util.config as config
-import foodflex.util.static as static
+import foodflex.data.firestore as data
+import foodflex.data.static as static
 import foodflex.periods.voting as voting
 import foodflex.periods.results as results
 import foodflex.periods.submissions as submissions
-import foodflex.periods.leaderboard as leaderboard
-
+import foodflex.data.leaderboard as leaderboard
 from foodflex.util.logging import logger
 
 # to use any command, the user must be in the admin list
 @bot.check
 async def is_admin(ctx):
+    
     authorised = ctx.author.id in config.admin_ids
     if not authorised:
         logger.warn(f'Unauthorised user \'{ctx.author.display_name}\' ({ctx.author.id}) tried to use command')
@@ -47,12 +45,10 @@ async def state(ctx):
         "participants": data.participants,
         "voting_map": data.voting_map,
         "period": data.period,
-        "mode": data.mode
     }
 
     as_json = json.dumps(state, indent=2)
     await ctx.send(f'```json\n{as_json}```')
-
 
 @bot.command(description='Switch to automatic time-based control')
 async def automatic(ctx):
@@ -60,19 +56,19 @@ async def automatic(ctx):
     data.set_mode('automatic')
 
 @bot.command(description='Start submissions period')
-async def submit(ctx):
+async def submission(ctx):
     await ctx.send('Manually switched to \'submissions\' mode')
     data.change_period('submissions', manual_change=True)
     await submissions.submission_period()
 
 @bot.command(description='Start voting period')
-async def vote(ctx):
+async def voting(ctx):
     await ctx.send('Manually switched to \'voting\' mode')
     data.change_period('voting', manual_change=True)
     await voting.voting_period()
 
 @bot.command(description='Start results period')
-async def result(ctx):
+async def results(ctx):
     await ctx.send('Manually switched to \'results\' mode')
     data.change_period('results', manual_change=True)
     await results.results_period()
